--- src/features/manage/components/manage-contact/manage-contact.tsx
@@ -1,3 +1,0 @@
-export const ManageContact = () => {
-  return <p>manage-contact</p>
-}
