--- src/features/manage/models/manage-models.ts
@@ -1,9 +1,0 @@
-export interface IManage {
-  title: string
-  category: string
-  price: number
-  photo: Array<string>
-  description: string
-  location: string
-  phone: string
-}
