--- src/features/manage/services/manage-util-service.ts
@@ -1,7 +1,0 @@
-const handlePublishProprety = (idProprety: string) => {
-  console.log('DELETE')
-}
-
-export const ManageUtilService = {
-  handlePublishProprety,
-}
