--- src/features/results/components/results-description/results-description.module.scss
@@ -1,74 +1,0 @@
-@import 'styles/convert.scss';
-
-.details {
-  display: flex;
-  justify-content: space-between;
-  gap: rem(30px);
-  margin: 24px 0 0 0;
-}
-
-.detailsText {
-  width: rem(750px);
-  height: rem(132px);
-  display: -webkit-box;
-  -webkit-line-clamp: 6;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  color: var(--gray-400);
-  padding: rem(12px) 0 0 0;
-}
-
-.detailsUser {
-  display: flex;
-  gap: rem(12px);
-  border: rem(1px) solid var(--gray-100);
-  border-radius: rem(12px);
-  width: rem(360px);
-  height: rem(142px);
-  padding: rem(20px);
-}
-
-.detailsAvatar {
-  height: rem(44px);
-  width: rem(44px);
-  border-radius: 50%;
-}
-
-.textColor {
-  display: flex;
-  flex-direction: column;
-  gap: rem(4px);
-}
-
-.titleColor {
-  color: var(--black);
-}
-
-.shareColor {
-  color: var(--primary-500);
-  font-size: rem(14px);
-}
-
-.detailsDescription {
-  padding: rem(32px) 0 0 0;
-  border-bottom: rem(1px) solid var(--gray-100);
-  border-top: rem(1px) solid var(--gray-100);
-}
-
-.shareIcon {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  cursor: pointer;
-}
-
-.detailsButton {
-  width: 100%;
-  display: flex;
-  padding: rem(24px) 0 0 0;
-  gap: 12px;
-}
-
-.purchaseButton {
-  width: rem(304px);
-}
